Refactor services to be in new modules logic.
